适配anan-boot-starter做升级2.7.0 控制层、服务层增加响应DTO支持 移除anan-platformapi模块，合并其代码到anan-cloudresoure模块

+++ anan-cloudadviced/anan-platformserver/src/main/java/top/fosin/anan/platform/dto/request/AnanInternationalCharsetCreateDto.java
@@ -21,11 +21,11 @@ public class AnanInternationalCharsetCreateDto implements Serializable {
 
     @NotNull(message = "国际化语言ID" + "{javax.validation.constraints.NotNull.message}")
     @ApiModelProperty(value = "国际化语言ID", example = "0")
-    private Integer internationalId;
+    private Long internationalId;
 
     @NotNull(message = "服务ID" + "{javax.validation.constraints.NotNull.message}")
     @ApiModelProperty(value = "服务ID", example = "0")
-    private Integer serviceId;
+    private Long serviceId;
 
     @ApiModelProperty(value = "自定义字符集", example = "String")
     private String charset;
